+++ notifications/CanceledEvent.php
@@ -51,12 +51,12 @@ class CanceledEvent extends BaseNotification
                 'contentTitle' => $this->getContentInfo($this->source, false),
                 'spaceName' =>  Html::encode($this->source->content->container->displayName),
             ]);
-        } else {
-            return Yii::t('ContentModule.notifications_views_ContentCreated', '{displayName} canceled the event "{contentTitle}".', [
-                'displayName' => Html::tag('strong', Html::encode($this->originator->displayName)),
-                'contentTitle' => $this->getContentInfo($this->source, false),
-            ]);
         }
+
+        return Yii::t('ContentModule.notifications_views_ContentCreated', '{displayName} canceled the event "{contentTitle}".', [
+            'displayName' => Html::tag('strong', Html::encode($this->originator->displayName)),
+            'contentTitle' => $this->getContentInfo($this->source, false),
+        ]);
     }
 
     /**
@@ -65,8 +65,8 @@ class CanceledEvent extends BaseNotification
     public function getMailSubject()
     {
         return Yii::t('CalendarModule.notification', '{displayName} canceled the event "{contentTitle}".', [
-            'displayName' => Html::encode($this->originator->displayName),
-            'contentTitle' => $this->getContentInfo($this->source, false),
+            'displayName' => $this->originator->displayName,
+            'contentTitle' => $this->getContentPlainTextInfo($this->source, false),
         ]);
     }
 }